Add type for controller methods and update env values
- Add SERVER_PORT, NODE_ENV on .env.local
- Update eslint config - disable `no-unused-vars` for *.d.ts files

+++ expressjs/src/models/setupModel.ts
@@ -3,18 +3,27 @@
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 
+import ENV from "@/constant/env";
 import knex from "@/models/knexConfig";
 import logger from "@/shared/utils/logger";
 import NavigationRepository from "./NavigationRepository";
 import UserRepository from "./UserRepository";
 
 const initModels = async () => {
+  if (ENV.NODE_ENV !== `local`) {
+    return;
+  }
+
   logger.log(`[MODEL] init models`);
   await UserRepository.createTable(knex);
   await NavigationRepository.createTable(knex);
 };
 
 const dropModels = async () => {
+  if (ENV.NODE_ENV !== `local`) {
+    return;
+  }
+
   logger.log(`[MODEL] drop models`);
   await UserRepository.dropTable(knex);
   await NavigationRepository.dropTable(knex);
